Basic Color Picking Widget (#39)
Used a Q_PROPERTY to implement the user property of the widget so that it can be mapped to models with a widget mapper. It handles transparency just fine using the transparency pattern and drawing a filled rectangle with the color on top of that. Went ahead and promoted the one in the room editor to it.