Add TreeModel::MakeResourcePath and ResourceMoved. (#199)
These should enable moving resources on disk in response to moving that resource in the tree.

We may need to revisit the TreeModel to make it use a custom wrapper of the TreeNode proto. This would save us some headache keeping track of the tree structure, and also where the files live on disk, when they were last saved, etc.