Add some utilities for facilitating drag-and-drop, and (partially) re-enable it.
It's broken right now. Very broken. Part of the problem is the GridView choosing bad coordinates (my guess is that the icon size is wrong). The other part of the problem is that this is allowed to disagree with the drop index being passed. A further problem is that the grid view completely disallows drop operations from inside itself when the computed index is past the bounds of the array. I have no idea why it was implemented this way.